Revert "Minor Fix"
This reverts commit a3e25f6a2361065642620a2471118ed86fb29c30.

Bundled Minor Changes

+++ Application.h
@@ -16,7 +16,7 @@ public:
 	void Init();
 	void Run();
 	void Exit();
-	static bool IsKeyPressed(int key);
+	static bool IsKeyPressed(unsigned short key);
 
 private:
